+++ include/game/minicomputer.hpp
@@ -41,17 +41,17 @@ struct miniScreenLineType {
     sfz::String     statusTrue;
     sfz::String     statusString;
     sfz::String     postString;
-    long            hiliteLeft;
-    long            hiliteRight;
-    long            whichButton;
+    int32_t         hiliteLeft;
+    int32_t         hiliteRight;
+    int32_t         whichButton;
     lineSelectType  selectable;
     bool         underline;
     lineKindType    lineKind;
-    long            value;      // for keeping track of changing values
-    long            statusType;
-    long            whichStatus;
-    long            statusPlayer;
-    long            negativeValue;
+    int32_t         value;      // for keeping track of changing values
+    int32_t         statusType;
+    int32_t         whichStatus;
+    int32_t         statusPlayer;
+    int32_t         negativeValue;
     baseObjectType* sourceData;
 };
 
@@ -63,23 +63,23 @@ void ClearMiniScreenLines( void);
 void ClearMiniObjectData( void);
 void draw_mini_screen();
 void MakeMiniScreenFromIndString( short);
-void MiniComputerHandleKeys( unsigned long, unsigned long);
-void MiniComputerHandleNull( long);
-long MiniComputerGetPriceOfCurrentSelection( void);
+void MiniComputerHandleKeys(uint32_t, uint32_t);
+void MiniComputerHandleNull(int32_t);
+int32_t MiniComputerGetPriceOfCurrentSelection( void);
 void UpdateMiniScreenLines( void);
 void draw_player_ammo(int32_t ammo_one, int32_t ammo_two, int32_t ammo_special);
 void draw_mini_ship_data(
         const spaceObjectType& newObject, unsigned char headerColor,
         short screenTop, short whichString);
 void MiniComputerDoAccept( void);
-void MiniComputerExecute( long, long, long);
+void MiniComputerExecute(int32_t, int32_t, int32_t);
 void MiniComputerDoCancel( void);
 void MiniComputerSetBuildStrings( void);
 void MiniComputerHandleClick( Point);
 void MiniComputerHandleDoubleClick( Point);
 void MiniComputerHandleMouseUp( Point);
 void MiniComputerHandleMouseStillDown( Point);
-void MiniComputer_SetScreenAndLineHack( long whichScreen, long whichLine);
+void MiniComputer_SetScreenAndLineHack(int32_t whichScreen, int32_t whichLine);
 
 }  // namespace antares
